requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
httpx
pypdf
python-pptx